Data model refactoring: ID of IdObject is now Long instead of Int
- applied all associated changes in listeners, controllers etc.
- fixed all existing unit tests

+++ sportstracker/src/main/java/de/saring/sportstracker/data/Entry.java
@@ -17,7 +17,7 @@ public abstract class Entry extends IdDateObject {
      *
      * @param id the ID of the object (null for new, not persisted objects)
      */
-    public Entry(Integer id) {
+    public Entry(Long id) {
         super(id);
     }
