+++ sportstracker/src/main/java/de/saring/sportstracker/gui/dialogs/SportTypeViewModel.java
@@ -23,7 +23,7 @@ import de.saring.util.data.IdObjectList;
  */
 public class SportTypeViewModel {
 
-    public final Integer id;
+    public final Long id;
     public final StringProperty name;
     public final BooleanProperty recordDistance;
     public final ObjectProperty<SpeedModeItem> speedMode;
